Bucket/requirements.txt
boto3
flask
python-dotenv
pyOpenSSL